requirements.txt
py-cord
dblpy
discord-components